% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixdist-package.R
\docType{package}
\name{mixdist-package}
\alias{mixdist}
\alias{mixdist-package}
\title{mixdist: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Gregor Laaha \email{gregor.laaha@boku.ac.at}

}
\keyword{internal}
