% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seasAM_FUN.R
\name{seasAM_FUN}
\alias{seasAM_FUN}
\title{Title}
\usage{
seasAM_FUN(x1)
}
\arguments{
\item{x1}{Vector of daily discharge series}
}
\value{
List containing annual series of AM, AMs, AMw
}
\description{
Title
}
