% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdfwei_Mixed_FUN.R
\name{cdfwei_Mixed}
\alias{cdfwei_Mixed}
\title{Mixed Weibull distribution}
\usage{
cdfwei_Mixed(x=AM, y=AM, para.x = pelwei(samlmu(AMs)), para.y = pelwei(samlmu(AMw)), plot=TRUE)
}
\arguments{
\item{x}{Vector of quantiles. Default is \code{AM}.}

\item{y}{Vector of quantiles. Default is \code{AM}.}

\item{para.x}{Numeric vector containing the parameters of the distribution.}

\item{para.y}{Numeric vector containing the parameters of the distribution.}

\item{plot}{Logical indicating if the function shall be added to a plot. Default is \code{FALSE}.}

\item{...}{Additional arguments.}
}
\value{
Vector of probabilities
}
\description{
Mixed Weibull distribution
}
