% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixdist-package.R
\docType{package}
\name{mixdist-package}
\alias{mixdist}
\alias{mixdist-package}
\title{mixdist: Mixed Distribution Estimators for Low Flows (Mixed & Mixed Copula)}
\description{
All functions for low-flow frequency analysis based on the two papers Laaha (2023a,b) in HESS .
}
\author{
\strong{Maintainer}: Gregor Laaha \email{gregor.laaha@boku.ac.at}

}
\keyword{internal}
